grr
I forgot to change the location of the header back to the relative thing
so I put that back so it would source the pic from the folder in here.
yay. but I checked on mobile and the fonts are messed up. I need to
upload tw cent somehow. the problem is that it's a font that's
automatically loaded onto windows so idk if I'll be able to find a ver
that we can use on a webpage. I'll look at google fonts and stuff ig.
idk if I can actually figure out how to direct it tho lmao